requirements.txt
flask
PyWechatAPI
qrcode
pillow